requirements.txt
crewai
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
qdrant-client
openai
anthropic
pyotp
python-jose[cryptography]
passlib[bcrypt]
sendgrid
twilio
langdetect
sentence-transformers
pytest
pytest-asyncio
python-dotenv
httpx
pydantic
alembic
